feat: add schemas for service orders including response and update models
